+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 9.2.2018.2
+Added Silent Mode toggle. Allows silently blocking players, entire realms, or everything. (Prevents whisper auto replies.)
+Fixed edge case where the dialog box for declining players could persist and force you to whisper someone.
+Added close button to above mentioned dialog box. (Note: This wont stop filtering them and the dialog box will return.)
+
 # 9.2.2018
 Fixed bad function call when unblocking target.
 Fixed sometimes having no realm when blocking a player out of render range. (untested with different than own realm.)